requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
discord==1.0.1
discord.py==1.6.0
distlib==0.3.1
Django==3.1.7
django-request==1.5.6
filelock==3.0.12
identify==2.1.3
idna==3.1
multidict==5.1.0
nodeenv==1.5.0
pre-commit==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
yarl==1.6.3

pytest~=6.2.3
